refactor: SiteUser에서 MyPage로 서비스명 변경하여 역할 명확화
